force windows parameters to be integers

+++ rcm/client/gui/rcm_main_window.py
@@ -57,12 +57,12 @@ class RCMMainWindow(QMainWindow):
         logger.debug("screen_width: " + str(screen_width))
         logger.debug("screen_height: " + str(screen_height))
 
-        self.setGeometry((screen_width / 2) - (app_width / 2),
-                         (screen_height / 2) - (app_height / 2),
-                         app_width, app_height)
+        self.setGeometry(int((screen_width / 2) - (app_width / 2)),
+                         int((screen_height / 2) - (app_height / 2)),
+                         int(app_width), int(app_height))
 
-        self.setMinimumHeight(app_height)
-        self.setMinimumWidth(app_width)
+        self.setMinimumHeight(int(app_height))
+        self.setMinimumWidth(int(app_width))
 
         self.build_menu()
